server_python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.28.0
Faker>=33.0.0
openai>=1.60.0
numpy>=2.0
tiktoken>=0.8.0
websockets>=14.0
langdetect>=1.0.9

# RAG Pipeline
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
langchain-text-splitters>=0.3.0
qdrant-client>=1.12.0
PyYAML>=6.0
